requirements.txt
Django==3.2.9
python-dotenv~=0.17.0

psycopg2-binary==2.8.6

pre-commit==2.11.1
reorder-python-imports~=2.6.0
black==21.7b0
flake8==3.9.0
flake8-django==1.1.1

djangorestframework~=3.12.0
djangorestframework-simplejwt~=4.7.2
django-filter~=21.1
drf-yasg~=1.20.0

gunicorn==20.1.0
whitenoise==5.2.0

